requirements.txt
Flask>=2.0.0
flask-cors>=3.0.0
transformers>=4.0.0
torch>=1.9.0
googletrans==4.0.0-rc1
requests>=2.25.0
python-dotenv>=0.19.0

# Additional dependencies
numpy>=1.19.0  # Required by transformers
tqdm>=4.65.0   # Required for progress bars
scipy>=1.7.0   # Required for audio processing
typing>=3.7.4  # For type hints
uuid>=1.30     # For generating unique IDs
gunicorn>=20.1.0  # For production deployment

# Development dependencies (optional)
pytest>=6.0.0  # For testing
pylint>=2.8.0  # For code quality
black>=21.0.0  # For code formatting